WRISTBAND-40 Allow MONGO_URI to be overriden from environment
We need to override this to connect to our replica set. We could add
an envvar for URI parameters but it's simpler to override the whole
URI.

Also make the `MONGO_DB_NAME` envvar the same in env and in config, which
is in line with other MongoDB libs like PyMongo.